+++ app/schemas/asset_responses.py
@@ -1,6 +1,9 @@
 from datetime import datetime
+
 from pydantic import BaseModel
-from src.types import AssetType
+
+from app.models import AssetType
+
 
 class BaseAssetResponse(BaseModel):
     asset_type: AssetType
@@ -9,14 +12,17 @@ class BaseAssetResponse(BaseModel):
     source: str
     cached_at: datetime
 
+
 class StockResponse(BaseAssetResponse):
     asset_type: AssetType = AssetType.STOCK
     full_name: str
     name: str
 
+
 class CryptoResponse(BaseAssetResponse):
     asset_type: AssetType = AssetType.CRYPTO
     name: str
+
 
 class SteamResponse(BaseAssetResponse):
     asset_type: AssetType = AssetType.STEAM