+++ src/types/response_enums.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-class AssetType(Enum):
-    STOCK = 'stock'
-    CRYPTO = 'crypto'
-    STEAM = 'steam'